Update lattice_paths.py
Will add explanatory comments for the code later.
This program is called "Speedy." and it is based on a genetic algorithm, which is way faster than the greedy and brute force algorithm.
For example, the  brute force algorithm would take months of computation for values of m greater than 5, where as Speedy may take less than a day as it found that there are in fact 14 k-distinct paths for a 6 by 3 lattice with k=5, and 19 k-distinct paths for a 7 by 3 lattice with k = 6 in less than one night. Better than the greedy, which found only 13 and 18 respectively after many weeks of computation.